requirements.txt
strands-agents[litellm]
strands-agents-tools
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
slack_sdk
